refactor: Split monitoring.py into focused command modules
Split the monolithic monitoring.py (1,267 LOC) into 6 focused modules,
one per command, to improve maintainability and reduce complexity.

Changes:
- monitoring_list.py (787 LOC) - list command + session helpers
- monitoring_ps.py (94 LOC) - ps command
- monitoring_session.py (158 LOC) - session command
- monitoring_status.py (107 LOC) - status command
- monitoring_top.py (129 LOC) - top command
- monitoring_w.py (95 LOC) - w command

Each module now has:
- Single command responsibility
- Clear public API via __all__
- Focused imports
- Module docstring with command description

Updated imports in:
- commands/__init__.py (import from new modules)
- cli.py (backward compatibility exports)
- commands/restore.py (get_vm_session_pairs)
- commands/sessions.py (get_vm_session_pairs)
- orchestrator.py (_get_config_int)
- tests/unit/test_bastion_tunnel_retry.py (all helpers)

Philosophy:
- One module = One command (single responsibility)
- Helpers stay with their primary command
- Maintain backward compatibility
- Clear public API boundaries

Part of Issue #423 - monitoring.py decomposition.
