+++ src/azlin/commands/monitoring_list.py
@@ -1,14 +1,9 @@
-"""Monitoring commands for azlin.
+"""LIST command for azlin monitoring.
 
-This module contains monitoring-related CLI commands extracted from cli.py.
-Part of Issue #423 - cli.py decomposition.
+This module contains the 'list' command (main VM listing logic) extracted from monitoring.py.
+Part of Issue #423 - monitoring.py decomposition.
 
-Commands:
-    - w: Run 'w' command on all VMs (who's logged in)
-    - ps: Run 'ps aux' command on all VMs (process listing)
-    - top: Distributed real-time monitoring dashboard
-    - status: Show VM status information
-    - session: Set or view session name for a VM
+Command:
     - list: List VMs in a resource group
 """
 
@@ -24,8 +19,6 @@ import click
 from rich.console import Console
 
 from azlin.azure_auth import AzureAuthenticator
-
-# Import list command helpers (refactored from monolithic list_command)
 from azlin.commands._list_helpers import (
     build_vm_table,
     configure_logging,
@@ -37,7 +30,6 @@ from azlin.commands._list_helpers import (
 )
 from azlin.config_manager import ConfigError, ConfigManager
 from azlin.context_manager import ContextError, ContextManager
-from azlin.distributed_top import DistributedTopError, DistributedTopExecutor
 from azlin.modules.bastion_detector import BastionDetector
 from azlin.modules.bastion_manager import BastionManager, BastionManagerError
 from azlin.modules.ssh_connector import SSHConfig
@@ -47,13 +39,13 @@ from azlin.network_security.bastion_connection_pool import (
     PooledTunnel,
     SecurityError,
 )
-from azlin.remote_exec import PSCommandExecutor, TmuxSession, TmuxSessionExecutor, WCommandExecutor
+from azlin.remote_exec import TmuxSession, TmuxSessionExecutor
 from azlin.tag_manager import TagManager
-from azlin.vm_manager import VMInfo, VMManager, VMManagerError
+from azlin.vm_manager import VMInfo, VMManagerError
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["list_command", "ps", "session_command", "status", "top", "w"]
+__all__ = ["get_vm_session_pairs", "list_command"]
 
 # SSH timeout configuration for tmux session detection
 # These values are based on empirical observation and conservative estimates:
@@ -61,478 +53,6 @@ __all__ = ["list_command", "ps", "session_command", "status", "top", "w"]
 # - Bastion: Routing through Azure Bastion adds ~5-7s latency, plus VM SSH startup ~3-5s, buffer to 15s
 DIRECT_SSH_TMUX_TIMEOUT = 5  # Seconds - Direct SSH connections (public IP)
 BASTION_TUNNEL_TMUX_TIMEOUT = 15  # Seconds - Bastion tunnels (routing latency + VM SSH startup)
-
-
-@click.command()
-@click.option("--resource-group", "--rg", help="Resource group", type=str)
-@click.option("--config", help="Config file path", type=click.Path())
-def w(resource_group: str | None, config: str | None):
-    """Run 'w' command on all VMs.
-
-    Shows who is logged in and what they are doing on each VM.
-
-    \b
-    Examples:
-        azlin w
-        azlin w --rg my-resource-group
-    """
-    try:
-        # Get resource group
-        rg = ConfigManager.get_resource_group(resource_group, config)
-
-        if not rg:
-            click.echo("Error: No resource group specified.", err=True)
-            sys.exit(1)
-
-        # Get SSH key
-        ssh_key_pair = SSHKeyManager.ensure_key_exists()
-
-        # List running VMs - use tag-based discovery (consistent with list command)
-        # This ensures azlin w detects same VMs as azlin list, including custom-named VMs
-        vms, _was_cached = TagManager.list_managed_vms(resource_group=rg, use_cache=False)
-        vms = [vm for vm in vms if vm.is_running()]  # Filter to running VMs
-
-        # Populate session names from tags (same logic as list command)
-        for vm in vms:
-            # Use tags already in memory instead of making N API calls
-            if vm.tags and TagManager.TAG_SESSION in vm.tags:
-                vm.session_name = vm.tags[TagManager.TAG_SESSION]
-            else:
-                # Fall back to config file
-                vm.session_name = ConfigManager.get_session_name(vm.name, config)
-
-        if not vms:
-            click.echo("No running VMs found.")
-            return
-
-        # Get SSH configs with bastion support (Issue #281 fix)
-        from azlin.cli_helpers import get_ssh_configs_for_vms
-
-        ssh_configs, routes = get_ssh_configs_for_vms(
-            vms=vms,
-            ssh_key_path=ssh_key_pair.private_path,
-            skip_interactive=True,  # Batch operation
-            show_summary=True,
-        )
-
-        if not ssh_configs:
-            click.echo("No reachable VMs found.")
-            return
-
-        click.echo(f"Running 'w' on {len(ssh_configs)} VMs...\n")
-
-        # Execute in parallel (bastion tunnels cleaned up automatically via atexit)
-        results = WCommandExecutor.execute_w_on_routes(routes, timeout=30)
-
-        # Display output
-        output = WCommandExecutor.format_w_output(results)
-        click.echo(output)
-
-    except VMManagerError as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except Exception as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-
-
-@click.command()
-@click.option("--resource-group", "--rg", help="Resource group", type=str)
-@click.option("--config", help="Config file path", type=click.Path())
-@click.option("--grouped", is_flag=True, help="Group output by VM instead of prefixing")
-def ps(resource_group: str | None, config: str | None, grouped: bool):
-    """Run 'ps aux' command on all VMs.
-
-    Shows running processes on each VM. Output is prefixed with [vm-name].
-    SSH processes are automatically filtered out.
-
-    \b
-    Examples:
-        azlin ps
-        azlin ps --rg my-resource-group
-        azlin ps --grouped
-    """
-    try:
-        # Get resource group
-        rg = ConfigManager.get_resource_group(resource_group, config)
-
-        if not rg:
-            click.echo("Error: No resource group specified.", err=True)
-            sys.exit(1)
-
-        # Get SSH key
-        ssh_key_pair = SSHKeyManager.ensure_key_exists()
-
-        # List running VMs - use tag-based discovery (consistent with list command)
-        # This ensures azlin ps detects same VMs as azlin list, including custom-named VMs
-        vms, _was_cached = TagManager.list_managed_vms(resource_group=rg, use_cache=False)
-        vms = [vm for vm in vms if vm.is_running()]  # Filter to running VMs
-
-        if not vms:
-            click.echo("No running VMs found.")
-            return
-
-        # Get SSH configs with bastion support (Issue #281 fix)
-        from azlin.cli_helpers import get_ssh_configs_for_vms
-
-        ssh_configs, _routes = get_ssh_configs_for_vms(
-            vms=vms,
-            ssh_key_path=ssh_key_pair.private_path,
-            skip_interactive=True,  # Batch operation
-            show_summary=True,
-        )
-
-        if not ssh_configs:
-            click.echo("No reachable VMs found.")
-            return
-
-        click.echo(f"Running 'ps aux' on {len(ssh_configs)} VMs...\n")
-
-        # Execute in parallel (bastion tunnels cleaned up automatically via atexit)
-
-        results = PSCommandExecutor.execute_ps_on_vms(ssh_configs, timeout=30)
-
-        # Display output
-        if grouped:
-            output = PSCommandExecutor.format_ps_output_grouped(results, filter_ssh=True)
-        else:
-            output = PSCommandExecutor.format_ps_output(results, filter_ssh=True)
-
-        click.echo(output)
-
-    except VMManagerError as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except Exception as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-
-
-@click.command()
-@click.option("--resource-group", "--rg", help="Resource group", type=str)
-@click.option("--config", help="Config file path", type=click.Path())
-@click.option(
-    "--interval",
-    "-i",
-    help="Refresh interval in seconds (default 10)",
-    type=int,
-    default=10,
-)
-@click.option(
-    "--timeout",
-    "-t",
-    help="SSH timeout per VM in seconds (default 5)",
-    type=int,
-    default=5,
-)
-def top(
-    resource_group: str | None,
-    config: str | None,
-    interval: int,
-    timeout: int,
-):
-    """Run distributed top command on all VMs.
-
-    Shows real-time CPU, memory, load, and top processes across all VMs
-    in a unified dashboard that updates every N seconds.
-
-    \b
-    Examples:
-        azlin top                    # Default: 10s refresh
-        azlin top -i 5               # 5 second refresh
-        azlin top --rg my-rg         # Specific resource group
-        azlin top -i 15 -t 10        # 15s refresh, 10s timeout
-
-    \b
-    Press Ctrl+C to exit the dashboard.
-    """
-    try:
-        # Get resource group
-        rg = ConfigManager.get_resource_group(resource_group, config)
-
-        if not rg:
-            click.echo("Error: No resource group specified.", err=True)
-            sys.exit(1)
-
-        # Get SSH key
-        ssh_key_pair = SSHKeyManager.ensure_key_exists()
-
-        # List running VMs - use tag-based discovery (consistent with list command)
-        # This ensures azlin top detects same VMs as azlin list, including custom-named VMs
-        vms, _was_cached = TagManager.list_managed_vms(resource_group=rg, use_cache=False)
-        vms = [vm for vm in vms if vm.is_running()]  # Filter to running VMs
-
-        if not vms:
-            click.echo("No running VMs found.")
-            return
-
-        # Get SSH configs with bastion support (Issue #281 fix)
-        from azlin.cli_helpers import get_ssh_configs_for_vms
-
-        ssh_configs, _routes = get_ssh_configs_for_vms(
-            vms=vms,
-            ssh_key_path=ssh_key_pair.private_path,
-            skip_interactive=True,  # Batch operation
-            show_summary=True,
-        )
-
-        if not ssh_configs:
-            click.echo("No reachable VMs found.")
-            return
-
-        click.echo(
-            f"Starting distributed top for {len(ssh_configs)} VMs "
-            f"(refresh: {interval}s, timeout: {timeout}s)..."
-        )
-        click.echo("Press Ctrl+C to exit.\n")
-
-        # Create and run executor (bastion tunnels cleaned up automatically via atexit)
-        executor = DistributedTopExecutor(
-            ssh_configs=ssh_configs,
-            interval=interval,
-            timeout=timeout,
-        )
-        executor.run_dashboard()
-
-    except VMManagerError as e:
-        # VMManagerError is already user-friendly
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except DistributedTopError as e:
-        # DistributedTopError is already user-friendly
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except KeyboardInterrupt:
-        click.echo("\nDashboard stopped by user.")
-        sys.exit(0)
-    except Exception as e:
-        # Log detailed error for debugging, show generic error to user
-        logger.debug(f"Unexpected error in distributed top: {e}", exc_info=True)
-        click.echo("Error: An unexpected error occurred. Run with --verbose for details.", err=True)
-        sys.exit(1)
-
-
-@click.command()
-@click.option("--resource-group", "--rg", help="Resource group", type=str)
-@click.option("--config", help="Config file path", type=click.Path())
-@click.option("--vm", help="Show status for specific VM only", type=str)
-def status(resource_group: str | None, config: str | None, vm: str | None):
-    """Show status of VMs in resource group.
-
-    Displays detailed status information including power state and IP addresses.
-
-    \b
-    Examples:
-        azlin status
-        azlin status --rg my-resource-group
-        azlin status --vm my-vm
-    """
-    try:
-        # Ensure Azure CLI subscription matches current context
-        try:
-            ContextManager.ensure_subscription_active(config)
-        except ContextError as e:
-            click.echo(f"Error: {e}", err=True)
-            sys.exit(1)
-
-        # Get resource group
-        rg = ConfigManager.get_resource_group(resource_group, config)
-
-        if not rg:
-            click.echo("Error: No resource group specified.", err=True)
-            sys.exit(1)
-
-        # List VMs - use TagManager to filter by managed-by=azlin tag
-        from azlin.tag_manager import TagManager
-
-        vms, was_cached = TagManager.list_managed_vms(resource_group=rg)
-
-        # Filter out stopped VMs by default (consistent with list command behavior)
-        # Note: list command doesn't filter by default but shows all,
-        # keeping include_stopped=True behavior for status
-
-        if vm:
-            # Filter to specific VM
-            vms = [v for v in vms if v.name == vm]
-            if not vms:
-                click.echo(f"Error: VM '{vm}' not found in resource group '{rg}'.", err=True)
-                sys.exit(1)
-
-        vms = VMManager.sort_by_created_time(vms)
-
-        if not vms:
-            click.echo("No VMs found.")
-            return
-
-        # Display status table
-        click.echo(f"\nVM Status in resource group: {rg}")
-        click.echo("=" * 100)
-        click.echo(f"{'NAME':<35} {'POWER STATE':<18} {'IP':<16} {'REGION':<15} {'SIZE':<15}")
-        click.echo("=" * 100)
-
-        for v in vms:
-            power_state = v.power_state if v.power_state else "Unknown"
-            # Display IP with type indicator (Issue #492)
-            ip = (
-                f"{v.public_ip} (Public)"
-                if v.public_ip
-                else f"{v.private_ip} (Private)"
-                if v.private_ip
-                else "N/A"
-            )
-            size = v.vm_size or "N/A"
-            location = v.location or "N/A"
-            click.echo(f"{v.name:<35} {power_state:<18} {ip:<16} {location:<15} {size:<15}")
-
-        click.echo("=" * 100)
-        click.echo(f"\nTotal: {len(vms)} VMs")
-
-        # Summary stats
-        running = sum(1 for v in vms if v.is_running())
-        stopped = len(vms) - running
-        click.echo(f"Running: {running}, Stopped/Deallocated: {stopped}\n")
-
-    except VMManagerError as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except Exception as e:
-        click.echo(f"Unexpected error: {e}", err=True)
-        sys.exit(1)
-
-
-@click.command(name="session")
-@click.argument("vm_name", type=str)
-@click.argument("session_name", type=str, required=False)
-@click.option("--resource-group", "--rg", help="Resource group", type=str)
-@click.option("--config", help="Config file path", type=click.Path())
-@click.option("--clear", is_flag=True, help="Clear session name")
-def session_command(
-    vm_name: str,
-    session_name: str | None,
-    resource_group: str | None,
-    config: str | None,
-    clear: bool,
-):
-    """Set or view session name for a VM.
-
-    Session names are labels that help you identify what you're working on.
-    They appear in the 'azlin list' output alongside the VM name.
-
-    \b
-    Examples:
-        # Set session name
-        azlin session azlin-vm-12345 my-project
-
-        # View current session name
-        azlin session azlin-vm-12345
-
-        # Clear session name
-        azlin session azlin-vm-12345 --clear
-    """
-    try:
-        # Resolve session name to VM name if applicable
-        resolved_vm_name = ConfigManager.get_vm_name_by_session(vm_name, config)
-        if resolved_vm_name:
-            vm_name = resolved_vm_name
-
-        # Get resource group
-        rg = ConfigManager.get_resource_group(resource_group, config)
-
-        if not rg:
-            click.echo("Error: No resource group specified and no default configured.", err=True)
-            click.echo("Use --resource-group or set default in ~/.azlin/config.toml", err=True)
-            sys.exit(1)
-
-        # Verify VM exists
-        vm = VMManager.get_vm(vm_name, rg)
-
-        if not vm:
-            click.echo(f"Error: VM '{vm_name}' not found in resource group '{rg}'.", err=True)
-            sys.exit(1)
-
-        # Clear session name
-        if clear:
-            cleared_tag = False
-            cleared_config = False
-
-            # Clear from tags
-            try:
-                cleared_tag = TagManager.delete_session_name(vm_name, rg)
-            except Exception as e:
-                logger.warning(f"Failed to clear session from tags: {e}")
-
-            # Clear from config
-            cleared_config = ConfigManager.delete_session_name(vm_name, config)
-
-            if cleared_tag or cleared_config:
-                locations = []
-                if cleared_tag:
-                    locations.append("VM tags")
-                if cleared_config:
-                    locations.append("local config")
-                click.echo(
-                    f"Cleared session name for VM '{vm_name}' from {' and '.join(locations)}"
-                )
-            else:
-                click.echo(f"No session name set for VM '{vm_name}'")
-            return
-
-        # View current session name (hybrid: tags first, config fallback)
-        if not session_name:
-            # Try tags first
-            current_name = TagManager.get_session_name(vm_name, rg)
-            source = "VM tags" if current_name else None
-
-            # Fall back to config
-            if not current_name:
-                current_name = ConfigManager.get_session_name(vm_name, config)
-                source = "local config" if current_name else None
-
-            if current_name:
-                click.echo(f"Session name for '{vm_name}': {current_name} (from {source})")
-            else:
-                click.echo(f"No session name set for VM '{vm_name}'")
-                click.echo(f"\nSet one with: azlin session {vm_name} <session_name>")
-            return
-
-        # Set session name (write to both tags and config)
-        success_tag = False
-        success_config = False
-
-        # Set in tags (primary)
-        try:
-            TagManager.set_session_name(vm_name, rg, session_name)
-            success_tag = True
-        except Exception as e:
-            logger.warning(f"Failed to set session in tags: {e}")
-            click.echo(f"Warning: Could not set session name in VM tags: {e}", err=True)
-
-        # Set in config (backward compatibility)
-        try:
-            ConfigManager.set_session_name(vm_name, session_name, config)
-            success_config = True
-        except Exception as e:
-            logger.warning(f"Failed to set session in config: {e}")
-
-        if success_tag or success_config:
-            locations = []
-            if success_tag:
-                locations.append("VM tags")
-            if success_config:
-                locations.append("local config")
-            click.echo(
-                f"Set session name for '{vm_name}' to '{session_name}' in {' and '.join(locations)}"
-            )
-        else:
-            click.echo("Error: Failed to set session name", err=True)
-            sys.exit(1)
-
-    except VMManagerError as e:
-        click.echo(f"Error: {e}", err=True)
-        sys.exit(1)
-    except ConfigError as e:
-        click.echo(f"Config error: {e}", err=True)
-        sys.exit(1)
 
 
 # ============================================================================
